refactor(app, medications page): Extract widgets
