+++ app/lib/common/pages/medications/widgets/module.dart
@@ -1,0 +1,3 @@
+export 'annotation_card.dart';
+export 'disclaimer.dart';
+export 'sub_header.dart';
